setup terrain world

+++ packages/utils/src/Types.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.8.24;
 
 struct VoxelCoord {
-  int32 x;
-  int32 y;
-  int32 z;
+  int16 x;
+  int16 y;
+  int16 z;
 }
